pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Bruno Dias", email = "bruno69correadias@gmail.com"},
]
dependencies = [
    "django>=4.2.7",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

